+++ docs/form_qc_coordinator/CHANGELOG.md
@@ -2,6 +2,11 @@
 
 All notable changes to this gear are documented in this file.
 
+## 1.4.2
+* Updates to support legacy ingest pipeline
+* Evaluates subsequent visits only for longitudinal modules
+* Rebuilt for event logging updates
+  
 ## 1.4.1
 * Fixes a bug in updating the visit error log