+++ docs/form_transformer/CHANGELOG.md
@@ -3,7 +3,10 @@
 All notable changes to this gear are documented in this file.
 
 ## 1.7.1
+* Updates to support legacy ingest pipeline
+* Rebuilt for ingest config updates
 * Bug fixes and improvements
+* Rebuilt for event logging updates
 
 ## 1.7.0
 * Reprocesses duplicate visit if failed to copy existing QC metadata